Keith (UI - Ongoing) 2022-12-10

+++ lib/models/menu_item_model.dart
@@ -6,7 +6,7 @@ class MenuItem {
   String menuItemName;
   double menuItemPrice;
   String menuItemDescription;
-  double menuItemOrder;
+  int menuItemOrder;
 
   MenuItem({
     required this.menuItemID,
